Fix database test execution
This fix requires checking an empty sqlite3 database into the repository.  This is less than ideal for several reasons,
among them that it changes every time it is regenerated (even if the schema has not changed).  It would be better if
this file could be generated by the test suite before it runs.