OpenCensus based tracing w/Jaeger
This is a spike for discussion, I'm aware the config won't merge with
Markus changes and will need rebasing.

Key things to note most important first:
 - the system degrades gracefully - if tracing is not configured, it
   has no overheads, if it is configured but the tracer is down the
   overhead is a log message every so often as the bundler fails to
   dispatch spans, but that is out of the request processing path
   so we're mostly paying some memory footprint, and the system
   load sheds spans and traces as well dynamically in order to
   prioritise availability.
 - we can add additional filters and logic in future, or switch to
   different exporting systems like commercial (e.g. honeycomb) or
   home-grown ones through the Exporter interface.
 - I've defaulted this off, except in curl, and there it is defaulting
   to a local Jaeger, so folk would have to opt-in to having it running
   in order to have any unexpected egress implications.

There's some additional work needed to make this better polished, such
as including cache identifiers in the RPC interceptor to automatically
identify which node a request was sent to, but as this is a spike for
discussion I will defer such work until we decide want this to go ahead.

There are examples in the client code of working with the system, but
in boils down to "Add a span where you want more detail, and annotate it
with the detail you want".