client.GetObject now streams content
I realised after implementing this that there is a comment asking
for implementations to look like HTTP - e.g. to have a ReadCloser
rather than the chan based interface I've used. I'm happy to add
a layer in for that - basically put a chan-to-ReadCloser adapter
on top - but I think the underlying system needs to be async
and not dependent on buffering entire files up.