Index the cached connection pool by cache pubkey
As far as I can tell we always know the public key so this should be
safe.

The test change is a result of the previous scheme causing a wrong
collision: the mock test had both caches on the same address and
so a cache hit occured in the client connection pool incorrectly.