Shutdown scheduler only after decrypting bundles
The previous scheduler code shut down the scheduler once the last
bundle's content block requests were triggered out to the caches, but we
still need the scheduler to operate until we have dealt with errors that
may occur after that point such as a cache that is unreachable or
returning corrupt content which would require new bundle requests.

To deal with the out of order and head-of-line blocking problem we will
roll through already received bundles in-process and just re-inject them
back into the appropriate channels rather than running multiple queues
and potentially missing out on other optimisation opportunities such as
early re-try of chunks from bad caches. The Enum should let us
discriminate cleanly between the cases.

BundleOutcomes are handled before any other scheduler actions, as
they are a control mechanism for the client.