Handle deferrals from the client in the scheduler
When the client is not ready to handle a fetch group,
it can now punt it back to the scheduler. This permits
the client to be largely ignorant of the buffering details and allows
the scheduler to have smarts about discarding those fetch groups in
future when cache health logic is added.