Get bundles from the middle of objects
This permits writing retry logic without it potentially having an entire
new here-to-end-of-object list of bundles being sent by the publisher;
we only want to ask for the chunks we have decided we are having trouble
with.

I have normalised everything to chunks and check chunk sizes are
consistent, if we do decide to make this vary, we can rework it fairly
easily; this will make the retry logic fairly painless.