Retry bundles from the client
This submits range limited bundle requests from the client rather than
bailing out immediately when a fault is detected. Further work is
required to deliver end to end working retries, in particular the
publisher needs a feedback channel to allow the client to request a
different set of caches for that bundle in some fashion.