Add client feedback on caches to the publisher
If the client cannot use a single cache it needs to be able to tell the
publisher not to use it in bundle generation. Under the normal end to
end model of the internet this is done by clients picking a different IP
address, or, by load balancers / reverse proxies that are authoritative;
in our 3-party model, this has to be negotiated explicitly.