Remove cache run queue
Using a queue to execute Cache RPC commands doesn't make a lot of sense - the
GRPC layer provides multiplexing and message fragmentation as needed
over the one HTTP channel. Removing the queue will allow downloads of
different objects to be using the same GRPC client without artificial
addition of latency once we add that capability to the client. Lastly
this fixes a bug where the Run goroutine never closed as we never closed
the backlog channel, which leaked cacheconns.