Don't notify the scheduler on chunk downloads
chunk downloads aren't the right trigger point for the scheduler for two
reasons - firstly there is a notification per cache, rather than one per
bundle, and secondly we're notifying the scheduler before we know
whether the bundle completes or not. Readahead of bundles should be
calibrated to deal with fully processed bundles instead.