Exchange L2 payments asynchronously
The L2 payments are not logically in critical path as noted by existing
XXX comments, but they do need to be executed as reliably as we can.
Putting them behind a worker goroutine in the cache connection allows
this cleanly.

The tracing context is preserved in a way that is new in this code base
but supported and intended by the Census library. Passing the context
around directly would require breaking the rules for static analysis
of context.Context - and is not necessary.

A further improvement would be to migrate the RPC to a stream, to reduce
RTT overheads between each ticket to the same cache (or we could spawn
more goroutines), but I'm leaving that for another day.

Even a localhost 5MB download test benefits from this modication:

x baseline
+ modified
+--------------------------------------------------+
|                +                                 |
|                +            x   x                |
|                +            *   x     x          |
|       +     +  +   x  +  +  *   x  x  x          |
|       +     +  +   x  *  *  *   *  *  x  x   x   |
|+      +     *  +   *  *  *  *   *  *  x  x   x  x|
|                        |_______AM_______|        |
|           |_________A_M______|                   |
+--------------------------------------------------+
    N           Min           Max        Median           Avg        Stddev
x  30           0.2          0.31          0.26         0.259   0.026826389
+  28          0.16          0.27          0.23    0.22321429   0.029319892
Difference at 95.0% confidence
        -0.0357857 +/- 0.0147668
        -13.8169% +/- 5.70146%
        (Student's t, pooled s = 0.0280563)