First attempt at a connection-tolerant logger
This logger is connection tolerant and fairly performant, it writes the
logs to a local journal and then, a background routine wakes up on the
interval/heartbeat and uploads any contents. Logs are automatically
rotated and on any error, retried.

There is a server in here but it is not complete, that will be addressed
in a future patch. The client and reader are well-tested at this point I
hope, however.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>