Metrics proxy server implementation
The broad shape in this patch is this:

- caches connect to our endpoint
- they can hammer it as much as they want; we don't care - we don't
  process what they send when they send it
- prometheus scrapes our ingress service to collect proxied cache
  metrics, at which point we render the metrics from the caches for
  prometheus.
- the metric scrapes from the caches are signed by the cache private key

We'll have TLS on our GRPC endpoint once fully deployed, so replay
attacks on the caches metrics are not a concern: a cache sending it's
own metrics twice is not a bother..

Caches can pretend to be more caches by creating more private keys - we
have no defense against at at this point, but when they stop sending
metrics, the metrics will age out.

Caches cannot cause us to store a higher sample rate, because we are
polling on the backend.

Caches cannot cause us to consume unlimited bandwidth and CPU - we won't
process what they send faster than 1 poll per 5seconds, and GRPC flow
control will choke their upload (and eventually h2 and then their TCP
window and eventually it looks like a regular DOS IP segment flood).

One in particular is that the current implementation is single-instance
- I'm fine with that today, simply because highly available for this
service is easily delivered by reconnecting to a new instance - but any
deeper sophistication we might want to put around DOS prevention for
this service would require additional capabilities. As whatever we do
with logs evolves I expect to come back to this service and evolve it in
an aligned fashion. It doesn't need the bucket mechanism today though,
because of the underlying differences in domain.