Support passing debug logging state to log client; rewrite constructor
The logger client needed to be able to set a "debug" flag to enable its
own debug logging. This happens way too early -- before the service
logger is created -- so therefore was controlled by an environment
variable which was objectionable.

This allows passing through flags into the options struct. The call into
the constructor for the base logging framework was rewritten to support
only one options has and avoid the multiple options form which was never
used prior to this change. Now, the options struct is directly filled
with defaults provided by the programs, providing the same functionality
with a little less fluff.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>