Permit secure connections to GRPC services
We want secure connections to our well known services - the publisher,
bootstrapd and the observability services. A first step is to make it
possible to configure whether things are secure or not. To avoid
breaking services this patch just enables the configuration, a follow on
patch will toggle the default.