Ledger block pagination support
For both block explorer and the corner case of more than (limit of one
request) blocks at the same height, being able to page through the
blocks in the ledger with a stable cursor is beneficial.

The current replication friendly format is extended with a stateless
token to permit pagination, with features to support the block explorer
such as reversing direction anywhere, including when off the end of the
result set.

Long term the block explorer should grow its own replica of the ledger,
to allow custom lookaside indices - such as a tree projection to
facilitate walking each branch and so on. However right now this feature
has little to no performance impact on the ledger, the ledger needs the
pagination support for complete replication support, and we want to
deliver the block explorer asap to allow developers to introspect the
ledger while also allowing other work to be done.