Hex encode all blockexplorer byte fields
Rather than go encoding of those fields, this is easier on the eyes.

The TransactionWitness could perhaps get a custom String method, buts it
not clear whether the meaning will always be constant or not, so to
avoid locking in meaning prematurely, be conservative here, and we can
discuss further later.