Blockchain store and replication client
This generalises the blockchain store and adds a reusable loosely
coupled replication client written to the existing GRPC API. Subscribers
are called back within a transaction and must successfully apply new
blocks or the transaction will be rolled back. Whether a particular
component should be a subscriber or not is a more complicated discussion
than this patch is about.

This patch includes a few discrete things that aren't worth breaking
out at this point but are worth identifying to make review easier.

* A new config parser helper to get durations.
* Consolidation of the existing chain stores from the publisher and the
  cache into one chain store within the ledger. The miner should also be
  consolidated eventually.
* A Replicator to do the replication
* The addition of a callback mechanism to the ChainDB to permit client
  side validation coupled to the chain - such as transaction validation
  - to be hooked in atomically.
* Migration of the TXID type from ledger/ to ledger/models to permit the
  use of TXID as a foreign column type in sqlboil without causing a
  circular package import cycle. This allows type safe code and less
  slice length checking.
* Rename ledger.ChainContext to Position
* Add docs to the `make gen` target.