Introduce Type and TypeInstance
We need a place to abstract out code generation questions for arrays,
pointers, and symbol names. We also need an abstract interface for
working with types.

This can be evolved to a more complete type system in future - e.g.
moving pointer and array knowledge onto a single instance and separating
out symbol information.