Generate fixed width array types
This includes arrays of structs and byte arrays.

Fixed width strings don't make a lot of sense: character encoding is
intrinsically variant as a single codepoint in UTF8 is not fixed in
serialisation width, and as strings are for human consumption the idea
of fixed-width strings seems likely. For now they have been excluded as
both complicated and unneeded.