Fix #260: remove MaxByteRange from Ranger
This involves making it possible to explicitly define the length or max
length of array-like types when they are contained within an array, and
the associated code generation tweaks. Some code cleanup presented
itself as part of making this change and I've done so, as keeping things
apart would have been more awkward.