cachecash-gcp: Set global configuration parameters
This allows setting global configuration parameters in the block
chain. These exist to allow blockchain and operational aspects of the
CDN to be updated in realtime without requiring code rolls across the
fleet, or hard forks in the block chain.

As the number of parameters should be fairly static over time, a code
generation approach has been used. This will provide editor support such
as code-completion and type safety around the use of configuration
parameters. Parameters are not permitted to change type over time, so
their content is validated in patches. In the event of unknown
parameters being present though, the cachecash-gcp tool will show the
content of unknown parameters as base64 encoded strings.

Closes #219

Review fixups